pyproject.toml
[project]
name = "sars-2-spike-domains"
version = "0.1.0"
description = "Some tests for SARS-CoV-2 S1 and S2 spike domains."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dark-matter>=7.0.4",
    "pytest>=9.0.1",
]

[project.scripts]
print-domains = "sars_2_spike_domains.cli.print_domains:main"

[tool.pytest.ini_options]
pythonpath = ["src"]

[build-system]
requires = ["uv_build>=0.8.17,<0.9.0"]
build-backend = "uv_build"
